requirements.txt
Django==5.1.4
djangorestframework==3.15.2
drf-spectacular==0.28.0